Move WooPayConnectIframe to ./woopay/connect directory
